Added enemy bullet script
Need to change the AI to face player when it shoots and to get the bullet facing in the correct direction

+++ Assets/Scripts/Collectables/AddHealth.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class AddHealth : MonoBehaviour
 {
-    [SerializeField] float addHealth;
+    [SerializeField] int addHealth;
     // Start is called before the first frame update
     void Start()
     {
